feat: add Services page with dynamic service listing and category selection
- Created Services.vue component for displaying a list of services with category filtering.
- Implemented services.js for managing service data and methods, including tracking events with Mixpanel.
- Enhanced user experience with responsive design and interactive elements for selecting service categories and viewing details.
